Move helidon version resolver to a standalone class
